--- src/main/kotlin/minesweeper/domain/Mine.kt
@@ -1,3 +1,0 @@
-package minesweeper.domain
-
-class Mine : Block
