--- src/main/kotlin/minesweeper/domain/None.kt
@@ -1,3 +1,0 @@
-package minesweeper.domain
-
-class None : Block
