+++ src/main/kotlin/minesweeper/domain/area/Area.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.area
 
 data class Area(val width: Width, val height: Height) {
