+++ src/main/kotlin/minesweeper/domain/block/Block.kt
@@ -1,0 +1,3 @@
+package minesweeper.domain.block
+
+interface Block
