+++ src/main/kotlin/minesweeper/domain/block/Mine.kt
@@ -1,0 +1,3 @@
+package minesweeper.domain.block
+
+class Mine : Block
