+++ src/main/kotlin/minesweeper/domain/block/None.kt
@@ -1,0 +1,3 @@
+package minesweeper.domain.block
+
+class None : Block
